Rename: 배너와 공지사항 패키지 분리
1. 배너 관련 파일은 resource/banner 패키지로 이동
2. 공지사항 관련 파일은 resource/notice 패키지로 이동
