+++ src/main/java/com/jandi/plan_backend/resource/banner/dto/BannerListDTO.java
@@ -1,12 +1,12 @@
-package com.jandi.plan_backend.resource.dto;
+package com.jandi.plan_backend.resource.banner.dto;
 
-import com.jandi.plan_backend.resource.entity.Banner;
+import com.jandi.plan_backend.resource.banner.entity.Banner;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class BannerRespDTO {
+public class BannerListDTO {
     private final Integer bannerId;
     private final LocalDateTime createdAt;
     private final String title;
@@ -14,7 +14,7 @@ public class BannerRespDTO {
     private final String linkUrl;
     private final String imageUrl;
 
-    public BannerRespDTO(Banner banner, String imageUrl) {
+    public BannerListDTO(Banner banner, String imageUrl) {
         this.bannerId = banner.getBannerId();
         this.createdAt = banner.getCreatedAt();
         this.title = banner.getTitle();
