+++ src/main/java/com/jandi/plan_backend/resource/notice/dto/NoticeListDTO.java
@@ -1,18 +1,18 @@
-package com.jandi.plan_backend.resource.dto;
+package com.jandi.plan_backend.resource.notice.dto;
 
-import com.jandi.plan_backend.resource.entity.Notice;
+import com.jandi.plan_backend.resource.notice.entity.Notice;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class NoticeRespDTO {
+public class NoticeListDTO {
     private final Integer noticeId;
     private final LocalDateTime createdAt;
     private final String title;
     private final String content;
 
-    public NoticeRespDTO(Notice notice) {
+    public NoticeListDTO(Notice notice) {
         this.noticeId = notice.getNoticeId();
         this.createdAt = notice.getCreatedAt();
         this.title = notice.getTitle();
